+++ app.js
@@ -13,6 +13,7 @@ const parksRouter = require('./routes/parks');
 const presidentsRouter = require('./routes/presidents');
 const parastatalsRouter = require('./routes/parastatals');
 const lakesRouter = require('./routes/lakes');
+const riversRouter = require('./routes/rivers');
 const searchRouter = require('./routes/search');
 const notFound = require('./middleware/notFound');
 const errorHandler = require('./middleware/errorHandler');
@@ -35,6 +36,7 @@ app.use('/api/v1/parks', parksRouter);
 app.use('/api/v1/presidents', presidentsRouter);
 app.use('/api/v1/parastatals', parastatalsRouter);
 app.use('/api/v1/lakes', lakesRouter);
+app.use('/api/v1/rivers', riversRouter);
 app.use('/api/v1/search', searchRouter);
 
 app.get('/', (req, res) => {